fix(web): render desktop auth handoff dynamically
