+++ apps/web/app/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata, Viewport } from 'next';
 import localFont from 'next/font/local';
+import Script from 'next/script';
 import React from 'react';
 import { APP_NAME, BASE_URL } from '@/constants/app';
 import './globals.css';
@@ -233,10 +234,12 @@ export default async function RootLayout({
       suppressHydrationWarning
     >
       <head suppressHydrationWarning>
-        {/* eslint-disable-next-line @next/next/no-sync-scripts -- Must run before React hydration; next/script nonce drift causes local E2E console errors. */}
-        <script src='/electron-runtime-init.js' />
-        {/* eslint-disable-next-line @next/next/no-sync-scripts -- Must run before React hydration; next/script nonce drift causes local E2E console errors. */}
-        <script src='/theme-init.js' />
+        <Script
+          nonce=''
+          src='/electron-runtime-init.js'
+          strategy='beforeInteractive'
+        />
+        <Script nonce='' src='/theme-init.js' strategy='beforeInteractive' />
       </head>
       <body className={bodyClassName}>
         {FlagBadgeProvider ? (
